# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MPI API Documentação",
        "description": "Documentação da API do sistema MPI",
        "contact": {
            "email": "luizleao@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/perfil": {
            "get": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Listar todos os perfils",
                "operationId": "3a8e08fdf84c78a7b023e3f21fac88cb",
                "responses": {
                    "200": {
                        "description": "Lista de perfils",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Perfil"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Criar um novo perfil",
                "operationId": "88d3e9750b7baa4f953acbc9744a13d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Perfil criado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/povo": {
            "get": {
                "tags": [
                    "Povos"
                ],
                "summary": "Listar todos os povos",
                "operationId": "d35fb68b6e2208270911da17a1c71dd2",
                "responses": {
                    "200": {
                        "description": "Lista de povos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Povo"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Povos"
                ],
                "summary": "Criar um novo povo",
                "operationId": "6b25409d62bcf8ea040744721be50c0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Departamento de TI"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Povo criado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/perfil/{id}": {
            "get": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Obter um perfil específico",
                "operationId": "0b283f3735d8343417ab322a0658e9bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do perfil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Perfil"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Atualizar um perfil específico",
                "operationId": "658c5fcf1be4d675af94be80dbbde48b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Excluir um perfil específico",
                "operationId": "62119cc878b8d06e9b9df5f0adeca5ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Perfil excluído"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/povo/{id}": {
            "get": {
                "tags": [
                    "Povos"
                ],
                "summary": "Obter um povo específico",
                "operationId": "2b17a675df1b9700e530c575ff9be376",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do povo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Povo"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Povos"
                ],
                "summary": "Atualizar um povo específico",
                "operationId": "9147bd6361eefd074b62e0f07f257ba8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Yanomami"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Povo atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Povos"
                ],
                "summary": "Excluir um povo específico",
                "operationId": "a6f92df10f3ec22cbc6f742bee6dcce4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Povo excluído"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Perfil": {
                "properties": {
                    "idPerfil": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Administrador"
                    }
                },
                "type": "object"
            },
            "Povo": {
                "properties": {
                    "idPovo": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "type": "string",
                        "example": "Tupi-guarani"
                    },
                    "codEtnia": {
                        "type": "string",
                        "example": "267.00"
                    },
                    "lingua": {
                        "type": "string",
                        "example": "Tupi-guarani"
                    },
                    "familia_linguistica": {
                        "type": "string",
                        "example": "Tupi"
                    },
                    "ufs_povos": {
                        "type": "string",
                        "example": "PA,MA"
                    },
                    "qtd_ti_povo": {
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}